+++ example/domain/word-guessing-game/mongodb/spring/dcb-autoconfig/src/main/kotlin/org/occurrent/example/domain/wordguessinggame/mongodb/spring/dcb/autoconfig/features/wordhint/RevealCharacterInWordHintAfterPlayerGuessedTheWrongWord.kt
@@ -16,13 +16,12 @@
 package org.occurrent.example.domain.wordguessinggame.mongodb.spring.dcb.autoconfig.features.wordhint
 
 import org.occurrent.annotation.Subscription
-import org.occurrent.application.converter.CloudEventConverter
 import org.occurrent.application.service.blocking.dcb.DcbApplicationService
 import org.occurrent.dsl.dcb.blocking.dcbPosition
 import org.occurrent.dsl.dcb.blocking.dcbTags
 import org.occurrent.dsl.dcb.blocking.queryForList
+import org.occurrent.dsl.query.blocking.DomainEventQueries
 import org.occurrent.dsl.subscription.blocking.EventMetadata
-import org.occurrent.eventstore.api.dcb.DcbEventStore
 import org.occurrent.example.domain.wordguessinggame.event.CharacterInWordHintWasRevealed
 import org.occurrent.example.domain.wordguessinggame.event.GameEvent
 import org.occurrent.example.domain.wordguessinggame.event.GameWasStarted
@@ -43,8 +42,7 @@ import kotlin.streams.asStream
 @Component
 class RevealCharacterInWordHintAfterPlayerGuessedTheWrongWord(
     private val applicationService: DcbApplicationService<GameEvent>,
-    private val eventStore: DcbEventStore,
-    private val cloudEventConverter: CloudEventConverter<GameEvent>
+    private val domainEventQueries: DomainEventQueries<GameEvent>
 ) {
 
     @Subscription(id = "WhenPlayerGuessedTheWrongWordThenRevealCharacterInWordHintPolicy")
@@ -59,8 +57,8 @@ class RevealCharacterInWordHintAfterPlayerGuessedTheWrongWord(
     @Retryable(backoff = Backoff(delay = 100, multiplier = 2.0, maxDelay = 1000))
     operator fun invoke(playerGuessedTheWrongWord: PlayerGuessedTheWrongWord) {
         val gameId = playerGuessedTheWrongWord.gameId
-        val wrongGuessCount = eventStore
-            .queryForList(GameDcbQueries.gameplay(gameId), cloudEventConverter)
+        val wrongGuessCount = domainEventQueries
+            .queryForList(GameDcbQueries.gameplay(gameId))
             .filterIsInstance<PlayerGuessedTheWrongWord>()
             .size